🗃️ ftree(web): move challenging filtes in the app folder out of  the (home) folder
