+++ apps/web/src/app/challenging/challenge/[challengeSlug]/page.tsx
