+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@codeEditor/default.tsx
