+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@codeEditor/page.tsx
