+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@tabContent/default.tsx
