+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@tabContent/result/default.tsx
