+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@tabContent/solutions/[solutionSlug]/default.tsx
