+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/@tabContent/solutions/default.tsx
