+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/default.tsx
