+++ apps/web/src/app/challenging/challenges/[challengeSlug]/challenge/layout.tsx
