+++ apps/web/src/app/challenging/challenges/[challengeSlug]/solution/[solutionSlug]/page.tsx
