+++ apps/web/src/app/challenging/challenges/[challengeSlug]/solution/page.tsx
